+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 All notable changes to this project will be documented in this file.
 See also [JsonSchemaChanges.md](JsonSchemaChanges.md) for a list of changes in the JSON Schema for API model files.
 
+## [ICD v1.3.2] - 
+
+### Added
+
+- Added "master" as an API version that is updated automatically from GitHub on startup or refresh of the web app.
+
+### Changed
+
+- Added new option to icd-git command: `--ingestMissing`: Ingests any APIs or ICDs that were published, but are not yet in the local database, plus any master branch versions.
+
+
 ## [ICD v1.3.1] - 2020-06-08
 
 ### Added